+++ Module.php
@@ -8,13 +8,10 @@
 namespace humhub\modules\popovervcard;
 
 use humhub\modules\popovervcard\models\Configuration;
-use Yii;
 use yii\helpers\Url;
 
 class Module extends \humhub\components\Module
 {
-
-
     public $resourcesPath = 'resources';
 
     /**
@@ -23,20 +20,9 @@ class Module extends \humhub\components\Module
     public function getConfigUrl()
     {
         return Url::to([
-            '/popover-vcard/admin'
+            '/popover-vcard/admin',
         ]);
     }
-
-    public function getName()
-    {
-        return Yii::t('PopoverVcardModule.base', 'VCard Popover');
-    }
-
-    public function getDescription()
-    {
-        return Yii::t('PopoverVcardModule.base', 'Shows a vcard popover in stream user and space links.');
-    }
-
 
     /**
      * Returns the module configuration model